The most recent version is "Final_version_kindanobosskill"
So basically it needs to be commented and the boss is still a mess. He
can kill you but it's pretty hard to kill him. The hitboxes on the
puffers are alright, but not perfect. I tried tweaking, but it wasn't
really working so they're there to stay. The game is alright. The boss
is a mess though, a huge mess. We might have to make him not kill you.